Implement vector division
The Physically Based Rendering book suggests calculating the reciprocal
of the divisor and multiplying by this since it is typically faster.

This makes the type signature a bit less generic because we need to be
able to cast the divisor to something that can be reciprocated.

For simplicity, we always return a Vector2f when dividing and we only
allow in-place division on Vector2f types.

We add the assert_approx_eq crate since floating point division is not
exact and the 3.0 / 10 assertion was failing.