Add a Ray struct with a constructor
In the C++ implementation, origin and direction
are passed as constant references. I think the
intention of this is to make it so no one can
modify this data. In Rust, we can instead move
ownership into Ray.

We haven’t built Medium yet, so add a dummy
implementation for now.